Remove broken tests and move constant files
