--- web/src/__tests__/setupTests.ts
@@ -1,3 +1,0 @@
-import fetchMock from 'jest-fetch-mock';
-
-fetchMock.enableMocks();
